+++ lib/market_hub.rb
@@ -30,6 +30,7 @@ require 'market_hub/api/mercado_livre/catalog'
 require 'market_hub/api/mercado_livre/variation'
 require 'market_hub/api/mercado_livre/question_answer'
 require 'market_hub/api/mercado_livre/order'
+require 'market_hub/api/mercado_livre/shipment'
 
 require 'market_hub/client'